added AddressServiceTest and UserServiceTest, added throwing exception when want to add or delete incorrect observed products

+++ backend/src/main/java/com/kkicks/backend/service/UserService.java
@@ -4,10 +4,8 @@ import com.kkicks.backend.config.security.JWTService;
 import com.kkicks.backend.dao.*;
 import com.kkicks.backend.entity.Address;
 import com.kkicks.backend.entity.Product.Product;
-import com.kkicks.backend.entity.ProductRating;
 import com.kkicks.backend.entity.User.Role;
 import com.kkicks.backend.entity.User.User;
-import com.kkicks.backend.entity.UserRating;
 import jakarta.persistence.EntityNotFoundException;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -23,10 +21,6 @@ public class UserService {
     @Autowired
     ProductDao productDao;
     @Autowired
-    ProductRatingDao productRatingDao;
-    @Autowired
-    UserRatingDao userRatingDao;
-    @Autowired
     AddressDao addressDao;
     public User createAdministrator(Long userId){
         User user = userDao.findById(userId).orElseThrow(() -> new EntityNotFoundException("user not found"));
@@ -37,7 +31,7 @@ public class UserService {
         return userDao.findAll();
     }
     public User findUserById(Long id){
-        return userDao.findById(id).orElse(null);
+        return userDao.findById(id).orElseThrow(() -> new EntityNotFoundException("user not found"));
     }
     public void deleteUserByID(Long userId){
         User user = userDao.findById(userId).orElseThrow(() -> new EntityNotFoundException("user not found"));
@@ -52,12 +46,18 @@ public class UserService {
     public User addProductToObserved(Long id, Long productId){
         User userToUpdate = userDao.findById(id).orElseThrow(() -> new EntityNotFoundException("User not found"));
         Product product = productDao.findById(productId).orElseThrow(() -> new EntityNotFoundException("Product not found"));
+        if(userToUpdate.getObservedProducts().contains(product)){
+            throw new IllegalArgumentException("User already observe this product!");
+        }
         userToUpdate.addProductToObserved(product);
         return userDao.save(userToUpdate);
     }
     public User deleteProductFromObserved(Long id, Long productId){
         User userToUpdate = userDao.findById(id).orElseThrow(() -> new EntityNotFoundException("User not found"));
         Product product = productDao.findById(productId).orElseThrow(() -> new EntityNotFoundException("Product not found"));
+        if (!userToUpdate.getObservedProducts().contains(product)){
+            throw new IllegalArgumentException("Product is not observed!");
+        }
         userToUpdate.deleteProductFromObserved(product);
         return userDao.save(userToUpdate);
     }
@@ -66,9 +66,9 @@ public class UserService {
         User userToUpdate = userDao.findById(id).orElseThrow(() -> new EntityNotFoundException("User not found"));
         userToUpdate.setFirstName(user.getFirstName());
         userToUpdate.setLastName(user.getLastName());
-        userToUpdate.setEmail(userToUpdate.getEmail());
-        userToUpdate.setUsername(userToUpdate.getUsername());
-        userToUpdate.setPassword(userToUpdate.getPassword());
+        userToUpdate.setEmail(user.getEmail());
+        userToUpdate.setUsername(user.getUsername());
+        userToUpdate.setPassword(user.getPassword());
         userToUpdate.setPhoneNumber(user.getPhoneNumber());
         userToUpdate.setAddress(user.getAddress());
         Address address = user.getAddress();
